Refactor helpers/user/current_user.go
